requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.0
pytesseract==0.3.10
pillow==10.0.0
opencv-python==4.8.0
pymupdf==1.23.0
librosa==0.10.0
soundfile==0.12.0
faiss-cpu==1.7.4
chromadb==0.4.10
pydantic==2.0.0
python-dotenv==1.0.0
requests==2.31.0
